VisualSizer Pro V2025.01.00
This is the first release of VisualSizer Pro with the new graphical user interface. Please be aware that there might still be some bugs in the program. We have tested thoroughly but need feedback from users to optimize functionality and correct where necessary. The motor database was copied from a previous version of VisualSizer and the data may eb out of date. This is again where we need feedback from users. However, the app comes with a database editor that allows to add, delete and modify motor data. For feedback, please refer to our website at https://visualsizer.com.